Move clipboard functions into own module and prettyprint JSON
